8/24 [sojoo] 토큰화 마무리
