--- .vscode/settings.json
@@ -1,6 +1,0 @@
-{
-    "files.associations": {
-        "builtin.h": "c",
-        "parse.h": "c"
-    }
-}
